test: extract shared InMemoryKeychain helper
Both `tests/unit/purge.test.ts` and `tests/integration/keychain-preload.ts`
defined their own copies of `InMemoryKeychain`. Move the implementation
to `tests/helpers/in-memory-keychain.ts` and import it from both sites so
any future change to the `KeychainBackend` interface has a single source
of truth to track.
