+++ tests/unit/purge.test.ts
@@ -8,34 +8,8 @@ import purgeCmd from '../../src/commands/purge';
 import { getDb, getDbPath, getFerretHome, resetDbCache } from '../../src/db/client';
 import { accounts, connections, rules, transactions } from '../../src/db/schema';
 import { ValidationError } from '../../src/lib/errors';
-import { type KeychainBackend, setKeychainBackend } from '../../src/services/keychain';
-
-class InMemoryKeychain implements KeychainBackend {
-  private store = new Map<string, string>();
-  private key(service: string, account: string): string {
-    return `${service}::${account}`;
-  }
-  async setPassword(service: string, account: string, password: string): Promise<void> {
-    this.store.set(this.key(service, account), password);
-  }
-  async getPassword(service: string, account: string): Promise<string | null> {
-    return this.store.get(this.key(service, account)) ?? null;
-  }
-  async deletePassword(service: string, account: string): Promise<boolean> {
-    return this.store.delete(this.key(service, account));
-  }
-  async findCredentials(service: string): Promise<Array<{ account: string; password: string }>> {
-    const prefix = `${service}::`;
-    const out: Array<{ account: string; password: string }> = [];
-    for (const [k, v] of this.store) {
-      if (k.startsWith(prefix)) out.push({ account: k.slice(prefix.length), password: v });
-    }
-    return out;
-  }
-  size(): number {
-    return this.store.size;
-  }
-}
+import { setKeychainBackend } from '../../src/services/keychain';
+import { InMemoryKeychain } from '../helpers/in-memory-keychain';
 
 function migrationsDir(): string {
   const here = fileURLToPath(new URL('.', import.meta.url));
